Make it possible to distinguish BSSL and AWS-LC (#458)
OpenSSL_version() can be a convenient method to distinguish AWS-LC from other projects at run-time. But it can't be used as a distinguisher from BSSL without changing the name to AWS-LC. Make this change since AWS-LC is a dependency in projects that can also depend on BSSL.

At the same time, modify the name of the project-specific header file (is_boringssl.h) to an AWS-LC-specific name.