Detect ACCP keys when parsing RSA private keys from ASN.1/DER (#436)
* Detect ACCP keys when parsing RSA private keys from ASN.1/DER

Prior to this change, all RSA private keys were also validated as public
keys during ASN.1 decoding. That was all well and good for CRT-enabled
private keys, as they contained the public exponent `e`, but was not
suitable for stripped private keys (i.e. keys defined only by `n` and
`d`) as required by the JCA and ACCP.

The refactor is based directly on @htorben's sketch [here][1], and does
not change AWS-LC's external API at all. We intentionally attempt to
minimize behavior and validation changes in order to minimize risk of
accidentally impacting RSA key generation and validation. To that end,
we only perform ACCP-aware validation when parsing from DER, and not
when doing key generation or general validation. The relevant RFC
([RFC-8107][2]) does not seem to explicitly permit omitted (i.e. NULL
type [ASN.1 tag value][3] of 5) RSA parameters when parsing RSA private
keys, so we consider an INTEGER value of 0 to indicate that a parameter
is missing in cases pertaining to ACCP.

This change resolves i/CryptoAlg-1026.

[1]: https://github.com/torben-hansen/aws-lc/compare/main...thoughts_rsa_accp_key_check
[2]: https://datatracker.ietf.org/doc/html/rfc8017#appendix-A.1.2
[3]: https://en.wikipedia.org/wiki/X.690#Identifier_octets

* Only check for NULL once parsed, detect ACCP via CRT params

This commit addresses feedback from @apetcher around having multiple
notions of "absent" in the same portions of RSA code. This commit
tightens that up by only considering `0` as "absent" in the DER/ASN.1
parsing code and only considering `NULL` as "absent" in the actual RSA
implementation code. See the test comment and prior commit message for
more detail around this.

Additionally, we now check _all_ RSA params (exponents, modulus and its
factors, and the CRT params) when detecting ACCP keys and adjust the
tests accordingly.

* Update crypto/rsa_extra/rsa_asn1.c

Co-authored-by: dkostic <25055813+dkostic@users.noreply.github.com>

* Refer to JCA instead ACCP

Co-authored-by: dkostic <25055813+dkostic@users.noreply.github.com>