Add Device Farm testing to Android CI dimension (#460)
Adds a bash and python script to let our Android CI dimension 
be able to upload and kick off the `AWSLCAndroidTestRunner` apk
tests in AWS Device Farm. With this change we are able to
confirm that our tests successfully pass in the Android
environment.

The python environment under `tests/ci/cdk` was moved to
`tests/ci` because the new script also relies on python
to kick off device farm with `boto3`.