Add curve25519 Montgomery ladder step
This Montgomery step is a typical combined implementation of doubling
and differential addition with swapping for curve25519, manipulating
points in projective coordinates without the Y coordinate. That is, a
pair (X,Z) is assumed to be a projective representation of affine point
(x,y) = (X / Z, y). If point = (X,Z) and pp = (n * (X,Z),[n+1] * (X,Z))
is a pair of points that are multiples of that point by n and n+1
respectively, then the output of either of the calls (which is more
efficient depends on platform)

  curve25519_ladderstep(rr,point,pp,b);

or

  curve25519_ladderstep_alt(rr,point,pp,b);

is a new pair rr = (n' * (X,Z),[n'+1] * (X,Z)) where n' = 2 * n + b, with
the input b considered to be a single 0 or 1 bit (though it's treated as a
C conditional so in fact any nonzero value means 1).