Move C++11 fwd-declared enum to ssl.h
AWS-LC's libcrypto library is standard C99, but there are a few
`#ifdef`'d forward-declared enums in `base.h`, which are exclusively
C++11 features. Further, they're C++11 features that are not supported
by GCC 4.1.2, which is the GCC installation on RHEL5. Since ACCP's
native library is C++, its CMake also tries to compile AWS-LC as C++,
triggering the `#if defined(__cplusplus)` macro and trying to compile
the forward-declared enums as C++. This fails on RHEL5 due to
aforementioned GCC version.

So, @htorben's solution is to simply move these forward declarations to
`ssl.h`, which is outside the object boundary of libcrypto. Both he and
I have confirmed that the resulting object compiles correctly with ACCP
under RHEL5 (RDE) environments.