Fix the static FIPS build on Linux (#496)
Fix the static FIPS build on Linux

The static FIPS build of AWS-LC was broken when the CPU capability
detection code was moved inside the FIPS module because the CPU
capability variables were included directly in `bcm.c`.  This change
fixes the issue by compiling the C file that contains the capability
variables (`cpucap.c`) separately from `bcm.c`, while retaining it in
the FIPS module. The commit also re-enables testing the static FIPS
build in the CI.

Co-authored-by: Dusan Kostic <dkostic@amazon.com>