Bump minimal cmake version to 3.0 (#538)
Bump minimal cmake version to 3.0.

To cater for non-unique executable names. First check whether cmake3 is present. If not, pick cmake as executable name. This assumes that cmake3 is actually at least version 3.0, but I don't know of any OS where this is not the case (covering all the OS's we claim to support).