add a sentinel |experimental_unstable_enable_sha3|
Use the sentinel value as a flag to enable and disable the SHA3 hashing algorithm. The default value is false - disabled. The user should use the public setter/getter to change/obtain the value of the sentinel. sha3_test.cc file implements testing for SHA3 by enabling the sentinel, then it is set back to disabled and the tests are repeated to ensure correct disabling of SHA3.