Fix CI timeouts
Some of the newer proofs (the worst case being arm/p521_jadd.ml) were
so slow they were hitting the CodeBuild timeout when run in CI, even
though they did all eventually work when run directly. This update
makes some crude speed improvements, mainly just applying cacheing to
core carry chain constructs that are repeated many times in macro
blocks but are still part of separate proof instances. It also changes
the order in which object files and proofs are handled to put the
point operations first (typically these are the more time-consuming
ones). But it also bumps up the CodeBuild timeout to the max anyway.
None of this changes the code or the verification results proved.