Added sanitizer specific images (#577)
The Sanitizer tests need access to `LLVMgold.so` in some situations in
order to properly perform the endianess checks. This change adds a new
Docker image that will have access to `LLVMgold.so` and is a
prerequisite to adding our endianess checks. This shouldn't change the
requirements for user builds of AWS-LC in any way.