Add curve25519 projective x-coordinate scalar multiplication
The function curve25519_pxscalarmul[_alt] takes the x-coordinate
of a point P on curve25519 and a 4-word (256-bit) scalar n, and
returns a projective representation (X,Z) of the x-coordinate of
n * P = (X/Z,Y). The point P can be (0,0) and the output will
still be correct. There is no specific encoding for the point
at infinity as input, but Z = 0 when that is the output.

This update also tweaks comments for curve25519_ladderstep[_alt]
to make explicit the assumption in the proof that the initial
point has Z = 1, i.e. is effectively the x coordinate of the
affine form of the point.