Link libjitterentropy into libcrypto.a (#511)
When building static FIPS build we were building a separate
`libjitterentropy.a` instead of linking it into `libcrypto.a`.
This hasn't caused any issues in our build and test system because
we had a `target_link_libraries` directive which instructs CMake
to always add `libjitterentropy.a` whenever `libcrypto.a` is linked
to another binary. However, if one tries to compile a program and link
it with `libcrypto.a` directly, the linker will complain about
missing symbol definitions from `libjitterentropy.a` and they will
have to link `libjitterentropy.a` as well. This should not be
a requirement, AWS-LC `libcrypto.a` should be a stand-alone library.
With this change we fix that.