allow static or shared libs to be used (#601)
Currently, a static lib that uses AWS-LC must use a static build of AWS-LC, it cannot use a shared build.

And vice-versa, a shared lib that uses AWS-LC must use a shared build of AWS-LC, it cannot use a static build.

With this change, either a static or shared build of AWS-LC can be used. If both are installed, then BUILD_SHARED_LIBS serves as a tie-breaker between the two.