Generalize field assumptions in curve25519 proofs so far
This changes only the formal specification and its proof, with the
code itself being unchanged. The previous version of the spec, given
the input X, was degenerate unless there was a corresponding Y in the
prime field F(2^255-19) with the affine point (X,Y) on the curve. But
the curve25519 X-coordinate functions like x25519 are explicitly
intended to work even on X inputs where the only suitable Y lives in
an extension field, and the new specification is general over all
fields of characteristic 2^255-19. This includes in particular the
quadratic extension F(2^255-19)[sqrt(2)], which has square roots for
all elements of the prime field and is used in Bernstein's original

  https://cr.yp.to/ecdh/curve25519-20060209.pdf