Add X25519 function
This provides the standard curve25519 "X25519" function including the
specific input mangling from "https://www.rfc-editor.org/rfc/rfc7748"
and proved correct against a spec based on Bernstein's original
description. The inputs and outputs are the usual s2n-bignum numbers
with four 64-bit word digits each. (That is, not arrays of bytes,
though on a little-endian machine with no special alignment
restrictions the difference amounts just to a pointer typecast.)