Fix proof success count
The embedded loads of the bignum_modinv.ml files in the new X25519
proofs resulted in their proof results being double-counted in the
final check for the "make proofs" target in the Makefile (both for ARM
and x86) and hence a toplevel failure even though all proofs worked.
Here we fix it by setifying the results to remove any duplicates, and
also correct two cases where ARM proofs re-used names instead of being
distinct (which was only shown up by this setification).