Add edwards25519 extended projective + precomputed mixed addition
This adds a point in extended-projective to a triple (y-x,x+y,2*d*x*y)
representing a point (x,y), intended for use in a setting where such a
triple is precomputed and stored in a table for efficient scalar
multiplication of a known point (e.g. a standard basepoint).

Also further README file extensions.