Add HKDF to ACVP Testing Harness (#658)
Currently AWS-LC has HKDF within the FIPS boundary, but doesn't support testing it with it's ACVP tool. This PR adds support for HKDF to the ACVP tool. Note that this does not add HKDF_expand() into the testing harness, as that requires it to be wrapped in an API (see CryptoAlg-1367).