Add X25519 basepoint function
The function "curve25519_x25519base" is the special case of
"curve25519_x25519" where the point X coordinate argument is the
standard basepoint 9. While faster than the general binary
function, it is functionally equivalent, i.e.

  curve25519_x25519base(res,n) =~= curve25519_x25519(res,n,9)

This update also adds support in the formal model for addressing
relative to the program counter (the ADR instruction for ARM, and
RIP-relative addresses for x86), which is used to index into the
tables in the code. In the current version of the code the
precomputed data tables are in the text segment directly after the
program (not in a separate .rodata segment), to ensures that the code
is position-independent and that the proof covers the actual code
without making assumptions about linker modifications.