Use command-line args in test/benchmark, extra TweetNaCl reference
This makes the testing and benchmarking program select what to
handle based on a rudimentary parse of the command-line arguments,
whereas previously the test program in particular used compile-time
selection. For example to benchmark only the "bignum_add" function
using 10000 inner repetitions:

 benchmark -10000 bignum_add

The output format of the testing in the case of anything but complete
success is also changed a bit to be more detailed and informative, but
the defaults, return codes and output on success are all the same.

This update also includes additional test runs of some of the
curve25519-related functions against an extract from TweetNaCl.