Fixing bug and adding a lower bound on input length.
- The input to the assembly functions is in bit length and not in byte length.
This still didn't solve the segfault because an input of 1 byte was
not dealt with properly.
- Added a lower bound to the input length to call the assembly
functions as is the case in the OpenSSL code.
This fixed the segfault but the vectors fail starting from the 5th
vector for both 128- and 256-bit for vectors in
crypto/cipher_extra/test/aes_<128|256>_gcm_tests.txt.