Fix the code signing issue on M1 Macs (#693) (#694)
On macOS 11 and higher on Apple Silicon, codesigning is mandatory for
binaries to run. This applies to both executables and dylibs. An ad-hoc
signature is sufficient, and the linker will automatically apply one when
a binary is created (see https://github.com/Homebrew/brew/issues/9082).

When we build libcrypto.dylib the linker automatically signs it. But then
we inject the FIPS integrity hash into libcrypto.dylib which changes the
binary so the signature applied by the linker is obviously not valid
anymore. So when an application, like crypto_test, tries to load
libcrypto.dylib it crashes because the signature is not valid. To work
round this we add an ad-hoc signature to `libcrypto.dylib` after the
FIPS integrity hash is injected.

Co-authored-by: dkostic <25055813+dkostic@users.noreply.github.com>