Deduplicate Kyber files (#705)
The commit removes the file duplication issue with Kyber implementation.
Previously, Kyber512 and Kyber768 implementations resided in separate
folders, both with a copy of all the source files. In this change we:
 - create a single _common_ folder with Kyber source code files,
 - add `kyber512_ref.c` file that includes the required C files
   from the common folder and selects the Kyber512 implementation
   by defining a macro (`KYBER_K 2`) specified in Kyber as the selector
   of the variant to be compiled.

In this way, when we add Kyber768 support we will only have to add the
corresponding `kyber768_ref.c` file, which will avoid duplicating the
Kyber source code files.