Return the number of bytes passed to the assembly function.
That's instead of returning the value returned from it,
which is 0, same as in OpenSSL when the number of
bytes passed is a multiple of 16.
(When the number of bytes passed in is 17, the return value
was 15).

In OpenSSL, the number of bytes passed to the function is
restricted to a number divisible by 16, denoted by the commented
part in the assignment of align_bytes.
They are commented because even when passing |len| bytes, AWS-LC
passes the same Wycheproof tests that are passed by OpenSSL and fails
the same ones.