Fix time measurement in Jitter on Apple platforms
On Apple platforms, Jitter RNG implementation uses `mach_absolute_time`
function to get the current value of the clock. However, this function
deals in "tick" units, not in time units. On Intel based Apple
platforms 1 tick = 1 ns, which is good enough timer resolution for
Jitter. On M1 macbooks 1 tick ~ 41.67 ns. This seems to be a problem
for Jitter since it causes frequent failures to produce enough entropy
on M1 macbooks. Another potential cause for the issue might be that
`mach_absolute_time` clock doesn't increment while they system is
asleep. I am not sure what is the definition of "asleep" on M1 macs,
how it correlates to performance and efficiency cores and different
power saving states, etc. So to be on the safe side we change:
  - Instead of `mach_absolute_time` we call `clock_gettime_nsec_np`
    function, as suggested by Apple's documentation [1],
  - Instead of using `CLOCK_UPTIME_RAW` that corresponds to
    `mach_absolut_time`, we use `CLOCK_MONOTONIC_RAW` that
    corresponds to `mach_continuous_time` whose clock increments
    even while the system is asleep [2].

[1] https://developer.apple.com/documentation/kernel/1462446-mach_absolute_time
[2] https://developer.apple.com/documentation/kernel/1646199-mach_continuous_time