Add first SM2 field operations, addition and subtraction
This begins the support of the SM2 curve from the "Commercial
Cryptography" standard (see GM/T 0003-2012 and ISO/IEC 14888). The
first two functions are addition and subtraction modulo the field
characteristic p_sm2 = 2^256 - 2^224 - 2^96 + 2^64 - 1.

These are highly analogous to P-256 functions, and in fact the ARM
bignum_add_p256 is tweaked here so that both adopt the same style,
using subtraction of the characteristic followed by selection instead
of masked resubtraction; this is slightly shorter and faster at the
expense of using more registers.