Fix FIPS static release build on ARM for gcc (#744)
This fixes the gcc release build for static FIPS on ARM. It
had been working on other compiler combinations for ARM, like
gcc debug or clang release/debug.

The following steps were done to fix this:
1. Fix three `delocate.peg` parsing issues.
2. An offset exists at the end of `:lo12:.LC9+8`, which
`delocate.go` was choking on. Removing the check gets around
this.
3. The gcc release build does not use `.comm` to define common
symbols. This creates a gap in the bss accessor functions, so we
use `bl` to catch these accessor function symbols instead.
4. Older versions of the gcc assembler maps variables of the
same value to each other, and skips label creation. 
`delocate.go` relies on these labels to determine if a symbol is
"known"; if a label doesn't exist, it's addressed at the end of
`bcm-delocated.S`. Rearranging the variable definition forces a
label creation and gets around the unwanted assembler
optimization.

Co-authored-by: Ubuntu <ubuntu@ip-172-31-32-154.us-west-2.compute.internal>